requirements.txt
# Core FastAPI Stack
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & Async
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.20.0  # Keep for local dev compatibility

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
email-validator==2.0.0

# AI Services (keep only what you're using)
google-genai==0.3.2  # More stable version
openai==1.3.0        # More stable version
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
websockets==12.0

# Remove these for production (not needed on Render):
# alembic>=1.16.4         # Remove - use your init_db.py instead
# watchdog>=3.0.0         # Remove - not needed in production
# celery>=5.5.3           # Remove unless you need async tasks
# redis>=6.4.0            # Remove unless using celery
# authlib>=1.6.2          # Remove unless using OAuth